feat: added in global samples controller which allows direct routing to samples just based on their ID and also replace Project::Samples::Metadata::FieldsController with Samples::MetadataController
