+++ app/views/samples/metadata/bulk_create.turbo_stream.erb
