+++ app/views/samples/metadata/bulk_update.turbo_stream.erb
