+++ config/routes.rb
@@ -20,6 +20,14 @@ Rails.application.routes.draw do
   scope path: '-' do
     resources :groups, only: %i[index new create]
     resources :projects, only: %i[index new create]
+    resources :samples, only: %i[show edit] do
+      resources :metadata, module: :samples, only: %i[edit update] do
+        collection do
+          post '/', action: :bulk_create
+          patch '/', action: :bulk_update
+        end
+      end
+    end
     resources :attachments, only: %i[show]
     draw :workflow_executions unless Irida::Pipelines.instance.available_pipelines.empty?
     resources :data_exports, only: %i[index new create destroy show] do
